Dev/wii/test

+++ backend/service/application/approvalWorkflowService.js
@@ -9,7 +9,7 @@ const approvalWorkflowSchema = require('../../validators/fabric/approvalWorkflow
 const AppError = require('../../utils/AppError');
 const approvalWorkflow = require('../fabric/approvalWorkflow');
 const submissionDao = require('../../dao/chaincodeMetadata/approvalWorkflowDao');
-const userDao = require('../../dao/userDao');
+const userDao = require('../../dao/user/userDao');
 const fileService = require('./fileService');
 
 class ValidationError extends Error {
@@ -79,13 +79,13 @@ class ApprovalWorkflowService {
         });
 
         // Pass firebase_uid to fabric service (for chaincode)
-        await approvalWorkflow.createSubmission({
-            submissionId,
-            owner: user?.uid,
-            role: user?.role,
-            proposalType,
-            docHash: fileMeta.docHash
-        });
+        // await approvalWorkflow.createSubmission({
+        //     submissionId,
+        //     owner: user?.uid,
+        //     role: user?.role,
+        //     proposalType,
+        //     docHash: fileMeta.docHash
+        // });
 
         return {
             submissionId,
@@ -104,16 +104,17 @@ class ApprovalWorkflowService {
             throw new AppError('Submission not found', 404, 'SUBMISSION_NOT_FOUND');
         }
 
-        await approvalWorkflow.deleteSubmission({
-            submissionId,
-            owner: user?.uid
-        });
-
-        if (metadata?.objectKey) {
+        if (metadata?.objectKey && metadata?.bucketName) {
             await fileService.deleteSubmissionFile({
+                bucketName: metadata.bucketName,
                 objectKey: metadata.objectKey
-            })
+            });
         }
+
+        // await approvalWorkflow.deleteSubmission({
+        //     submissionId,
+        //     owner: user?.uid
+        // });
 
         // Copilot note: Use metadata.owner (which is user_id from database) for DAO deletion
         await submissionDao.deleteSubmission({
@@ -130,10 +131,10 @@ class ApprovalWorkflowService {
         const validated = this.validate('submitForApproval', { params });
         const submissionId = validated.params.submissionId;
 
-        return await approvalWorkflow.submitForApproval({
-            submissionId,
-            owner: user?.uid
-        });
+        // return await approvalWorkflow.submitForApproval({
+        //     submissionId,
+        //     owner: user?.uid
+        // });
     }
 
     async approveSubmission({ params, body, user }) {
@@ -141,11 +142,11 @@ class ApprovalWorkflowService {
         const submissionId = validated.params.submissionId;
         const { remarks } = validated.body;
 
-        return await approvalWorkflow.approveSubmission({
-            submissionId,
-            approver: user?.uid,
-            remarks
-        });
+        // return await approvalWorkflow.approveSubmission({
+        //     submissionId,
+        //     approver: user?.uid,
+        //     remarks
+        // });
     }
 
     async requestChanges({ params, body, user }) {
@@ -153,11 +154,11 @@ class ApprovalWorkflowService {
         const submissionId = validated.params.submissionId;
         const { remarks } = validated.body;
 
-        return await approvalWorkflow.requestChanges({
-            submissionId,
-            approver: user?.uid,
-            remarks
-        });
+        // return await approvalWorkflow.requestChanges({
+        //     submissionId,
+        //     approver: user?.uid,
+        //     remarks
+        // });
     }
 
     async rejectSubmission({ params, body, user }) {
@@ -165,11 +166,11 @@ class ApprovalWorkflowService {
         const submissionId = validated.params.submissionId;
         const { remarks } = validated.body;
 
-        return await approvalWorkflow.rejectSubmission({
-            submissionId,
-            approver: user?.uid,
-            remarks
-        });
+        // return await approvalWorkflow.rejectSubmission({
+        //     submissionId,
+        //     approver: user?.uid,
+        //     remarks
+        // });
     }
 
     async resubmitSubmission({ params, body, user, file }) {
@@ -193,11 +194,11 @@ class ApprovalWorkflowService {
             size: fileMeta.size
         });
 
-        await approvalWorkflow.resubmitSubmission({
-            submissionId,
-            owner: user?.uid,
-            newDocHash: fileMeta.docHash
-        });
+        // await approvalWorkflow.resubmitSubmission({
+        //     submissionId,
+        //     owner: user?.uid,
+        //     newDocHash: fileMeta.docHash
+        // });
 
         return {
             submissionId,
@@ -218,9 +219,9 @@ class ApprovalWorkflowService {
         const validated = this.validate('getSubmissionHistory', { params });
         const submissionId = validated.params.submissionId;
 
-        return await approvalWorkflow.getSubmissionHistory({
-            submissionId
-        });
+        // return await approvalWorkflow.getSubmissionHistory({
+        //     submissionId
+        // });
     }
 }
